requirements.txt
pyvisa
pyserial
keithley2600
propar
pyyaml